feat(copy-link): add copying links to mobile tab
